Add game states.
Major refactoring of how game states work, leaves room for making the
states more advanced which was getting tedious.

Also fixes and closes issue #3, if the window document does not have
focus we fallback to set timeout, which will run at a slower interval
but keep the game from going into what is effectively an infinite loop
as far as the browser is concerned.